Test config refactoring for msg and token config (#768)
## Motivation


## Solution
Config Clean up for msg details and token config in test
Categorize different msg type - Data, Token, DataWithToken
Simplifying method signature for `(lane *CCIPLane) DeployNewCCIPLane`
Updating config to specify `NoOfTokensWithDynamicPrice` out of set of
available tokens per chain

+++ integration-tests/ccip-tests/testreporters/ccip.go
@@ -47,7 +47,7 @@ type TransactionStats struct {
 	GasUsed            uint64 `json:"gas_used,omitempty"`
 	TxHash             string `json:"tx_hash,omitempty"`
 	NoOfTokensSent     int    `json:"no_of_tokens_sent,omitempty"`
-	MessageBytesLength int    `json:"message_bytes_length,omitempty"`
+	MessageBytesLength int64  `json:"message_bytes_length,omitempty"`
 	FinalizedByBlock   string `json:"finalized_block_num,omitempty"`
 	FinalizedAt        string `json:"finalized_at,omitempty"`
 	CommitRoot         string `json:"commit_root,omitempty"`
